+++ rust/cyclotron-fetch/src/fetch.rs
@@ -1,11 +1,7 @@
 use std::{cmp::min, collections::HashMap, sync::Arc};
 
 use chrono::{DateTime, Duration, Utc};
-use cyclotron_core::{
-    base_ops::{Job, JobState},
-    error::QueueError,
-    worker::Worker,
-};
+use cyclotron_core::{Job, JobState, QueueError, Worker};
 use futures::StreamExt;
 use http::StatusCode;
 use reqwest::Response;
